Use Python 3 division behavior on Python 2

+++ master/buildbot/test/unit/test_data_logchunks.py
@@ -12,6 +12,9 @@
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
 # Copyright Buildbot Team Members
+from __future__ import division
+from __future__ import print_function
+
 import textwrap
 
 from future.builtins import range
@@ -93,7 +96,7 @@ class LogChunkEndpointBase(endpoint.EndpointMixin, unittest.TestCase):
                              {'logid': logid, 'firstline': i, 'content': expLines[i] + '\n'})
 
         # half and half
-        mid = len(expLines) / 2
+        mid = int(len(expLines) / 2)
         for f, l in (0, mid), (mid, len(expLines) - 1):
             logchunk = yield self.callGet(path,
                                           resultSpec=resultspec.ResultSpec(offset=f, limit=l - f + 1))
